# docs/swagger.yaml
basePath: /api/v1
definitions:
  database.Deployment:
    properties:
      deployedAt:
        type: string
      environment:
        type: string
      id:
        type: integer
      serviceID:
        type: integer
      status:
        type: string
      version:
        type: string
    type: object
  database.Project:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      tenantID:
        type: integer
      updatedAt:
        type: string
    type: object
  database.Service:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      language:
        type: string
      name:
        type: string
      projectID:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
    type: object
  database.Tenat:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  main.createDeploymentRequest:
    properties:
      environment:
        example: aws
        type: string
      status:
        example: in test
        type: string
      version:
        example: 1.0.0
        type: string
    type: object
  main.createProjectRequest:
    properties:
      description:
        example: Customer-facing portal
        type: string
      name:
        example: Tenant Portal
        type: string
    type: object
  main.createServiceRequest:
    properties:
      language:
        example: go
        type: string
      name:
        example: Auth Service
        type: string
      type:
        example: backend
        type: string
    type: object
  main.createTenantRequest:
    properties:
      email:
        example: admin@acme.com
        type: string
      name:
        example: Acme Corp
        type: string
    type: object
  main.errorResponse:
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
info:
  contact: {}
  description: REST API for tenants, projects, services, and deployments.
  title: Tenants API
  version: "1.0"
paths:
  /projects/{id}/services:
    get:
      description: Get all services for a project id
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Service'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: List services by project
      tags:
      - Services
    post:
      consumes:
      - application/json
      description: Create a new service for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Service payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.createServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Service'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Create service
      tags:
      - Services
  /services/{id}/deployments:
    get:
      description: Get all deployments for a service id
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Deployment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: List deployments by service
      tags:
      - Deployments
    post:
      consumes:
      - application/json
      description: Create a new deployment for a service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      - description: Deployment payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.createDeploymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Deployment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Create deployment
      tags:
      - Deployments
  /tenants:
    get:
      description: Get all tenants
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Tenat'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: List tenants
      tags:
      - Tenants
    post:
      consumes:
      - application/json
      description: Create a new tenant
      parameters:
      - description: Tenant payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.createTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Tenat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Create tenant
      tags:
      - Tenants
  /tenants/{id}:
    get:
      description: Get a tenant by tenant id
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Tenat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Get tenant by id
      tags:
      - Tenants
  /tenants/{id}/projects:
    get:
      description: Get all projects for a tenant id
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Project'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: List projects by tenant
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a new project for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.createProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Create project
      tags:
      - Projects
schemes:
- http
swagger: "2.0"
tags:
- description: Tenant management endpoints.
  name: Tenants
- description: Project management endpoints.
  name: Projects
- description: Service management endpoints.
  name: Services
- description: Deployment management endpoints.
  name: Deployments